Most events can now happen throughout the day.  This includes jobs, construction, base discovery, research, and events.
This new system also changed the non-research base tasks.  Construction has been replaced with a CPU Pool, which works on maintenance, construction, and jobs, in that order.  Bases not assigned to any task will also contribute to the CPU Pool, they must be assigned to Sleep (a new task) to remain idle and get the bonus against discovery.