Use XDG_CONFIG_DIR/singularity for preferences dir
Make E:S compliant to the XDG Base Directory Specification, helping
unclutter the user's $HOME of dotfiles/dirs. Also change default dir name
from "endgame" to "singularity", to be uniform with game's package name
in Arch, Debian, Ubuntu (and also with game's launcher, singularity.py)

This is backwards-compatible with previous releases: if ~/.endgame is found,
it is used. The new default is, for most scenarios, ~/.config/singularity