add new hotkey engine
For button hotkeys, embed the hotkey in the translatable string itself
using an & marker, and replace hotkey='x' for autohotkey=True, as in this
example: button.Button(... , text=_("FINANC&E"), autohotkey=True, ...)

A helper function used by Button class scans the translated string (which
will hopefully also contain an & marker, as in "&VOLTAR"), strips the text
of the & marker (for the label) and assigns the marked character as the
button's hotkey. Old usage of manually assigning hotkey="char" in button
definition still works, and will override autohotkey=True.

While Button class automatically handles hotkeys, g.hotkey(string)
function can be used anywhere else. It returns a dict with several info
about text and hotkeys (like char, position, stripped text, etc). More
details in source code docstring.

Note that this engine is technically independent of i18n: _() itself is
completely unaware of embedded hotkeys. For _(), "&BACK" is a string like
any other: it searches the PO file for this entry and returns the matching
string.

As with i18n, a current limitations is that location hotkeys don't work
properly when language is changed on the fly. This was always a bug, but
now it is much more evident,