improve data dir handling
Set the data directory (previously data_loc, now data_dir) in a way that
is not dependent on current directory. This means E:S may be invoked from
any anywhere, and using the linux launcher to change to game dir is not
necessary anymore.

So it is possible to start the game as:
python path/to/game/wherever/it/is/singularity.py

This also may fix setup.py not working in linux, although it wasn't tested

Also handle trailing "/" and absolute/relative path in all uses of data_dir