add new tool: util/gettext-singularity, customized gettext tools wrapper
Allow the creation and maintence of translation files, customized
for E:S. It automatically the following GNU's gettext tools:
- xgettext : to create or update data/messages.pot template
- msgmerge : to update existing data/messages_ll_CC.po files
- msgfmt   : to check existing PO files for syntax errors
- msginit  : to create a new data/messages_ll_CC.po file

For new files, also ask user the name of the language in native language
(eg, "Português Brasileiro" for pt_BR). This information can be used in
the future in Options Screen

execute util/gettext-singularity --help for more info.