load messages as soon as language is changed
Allow strings to be translated to chosen language very early in game:
in the preferences file parser error messages and in command-line
argument handling, including the --help page.

Previously the only languages available at that time were the game's
default (en_US) and the user's native locale. Settings in the preferences
file or in command line were not effective until main screen was loaded.

As a welcome side effect, messages are automatically loaded when setting
language, so load_messages() does not need to be manually invoked anymore.
set_language() will handle all locale and language changes.