update maketree so it works again
Also now it can be invoked from anywhere. It generate 4 files in current
directory: 2 .dot files and 2 png images, for techs and items, a tree
diagram of all technologies, pre-requisites and unlocked items.

It still has bugs and room for several improvements, but since it is
fully working there's no reason to hold it back, so I cherry-picked it
from my testing/classdiagram branch to master.

It is a very powerful tool, and its diagrams are invaluable for anyone
fiddling with game balance. A truly golden gem from the past just brought
back to life.  Thanks aes and FM!