properly handle data types of option vars (Closes: #132)
Previous assignment of gg.fullscreen = pygame.FULLSCREEN
caused prefs.dat file to get currupted, since pygame.FULLSCREEN
is a negative integer while fullscreen option is expected to
be boolean, outputting "Invalid or missing 'fullscreen' setting
in preferences." and turning fullscreen OFF when it was supposed
to be ON.

Proper handling was also applied to g.nosound, which is now a real
boolean throughout the code.

And, just to be safe, type convertion was applied to all vars
when writing the preferences file.