account for panels and title bar when resizing main window
When changing resolution in windowed mode to one of the standard resolutions,
make window a little smaller than requested, currently (-70,-70) pixels, to
account for desktop panels and title bar. This way a requested window of
1024x768 will fit nicely in a 1024x768 desktop.

This margin is not applied when selecting a custom resolution or when running
in fullscreen mode.