be more restrictive (and smarter) when applying desktop margins
Margins for windowed mode are now applied only if screen size matches
desktop size, as it makes little sense to apply margin for a window
already smaller than desktop.

It also sets margins for width and height independently, so only the
needed dimension is adjusted.

For example, in a desktop of 1366x768:
- A 1366x768 window is reduced in both width and weight
- A 1024x768 window is reduced in height only
- A  800x600 window is not reduced at all