player.py: Consitently use "self" instead of "g.pl"
We expect there to be exactly one Player object.  This implies that
inside the Player class "g.pl" and "self" will be the same object.
However, the code interleaves usage of "g.pl" and "self" despite it
being the same object.  This commit rectifies that by replacing all
uses of "g.pl" with "self" in the Player class.

Signed-off-by: Niels Thykier <niels@thykier.net>