Let listbox handle double-click events
Previously, dialog's would have to register the double-click event and
figure out if the user double-clicked on an item.  However, it seems
odd to have the parent handle something the listbox can do itself.
Plus the dialog only checked that the double-click happened over the
listbox rather than the user actually double-clicking an item.

With the new change, the listbox now exposes a
"on_double_click_on_item" kwarg parameter, which can be used to
register a reaction to the double-click event on the item.

Signed-off-by: Niels Thykier <niels@thykier.net>