Rewrite loading of techs to support adding new techs
Previously, when we load a savegame, the program would assume that the
savegame contained all the *current* technologies.  However, this
assumption fails every time a new technology is added to singularity
and then singularity simply crashes.

This patch solves this by pulling all techs from the data files and
then merging the few relevant bits ("done" and "cost_paid" from the
savegame) into the existing saves. Obsolete techs (i.e. techs prevent
in the savegame but not in the data files) are ignored to avoid
crashing on them.  In the case, where a technology is renamed, we can
preserve it by correcting it's "id" in the convert_from method.

With this patch, a save game made on commit b16134996 can still be
successfully loaded in 03b0bd51e8, where a new tech is introduced.

Signed-off-by: Niels Thykier <niels@thykier.net>