Use None as base value for {prev,last}_discovery instead of str
This avoid confusing type checks/IDEs as they think these fields
contain strings when they in fact hold locations (and only use "" as a
replacement for None).

As our __cmp__ is not None-safe, we implement an __eq__ that is.  This
also remove need for supporting str/unicode in the __cmp__ method.

Signed-off-by: Niels Thykier <niels@thykier.net>