Change the meaning of the extra values in {cash,cpu}_display
Previously, we have been using "future_cash" for cash_display as a
form of "current cash - remaining cost of (active) buyables".  This
has been replaced with an estimate of the money difference in the next
24 hours.

For the cpu_display, we previously used the current size of the CPU
pool.  This has been replaced by the number estimating the difference
between available CPU pool and tasks needing CPU.  Negative implies
that the tasks are missing CPU time while positive means some CPU
time is used for jobs.

Closes: #21
Signed-off-by: Niels Thykier <niels@thykier.net>