Avoid gratious warnings about the CPU pool
Previously, we would emit warnings based on the size of the explicitly
allocated CPU pool.  However, the game uses the effective CPU pool
(i.e.  including the unallocated CPU as well).  With this commit, the
warning also use the effective CPU pool rather than the allocated CPU
pool to make the warning match the game mechanics.

Signed-off-by: Niels Thykier <niels@thykier.net>