Recalculate CPU usage after changing sleep/wake state of a base
Fixes a regression from 5ff71b1. When the sleep/active state of a base
changed, the available CPU didn't update immediately. This looks like
the logical place to add it if 5ff71b1 is correct, though unless
MapScreen.rebuild is called on every frame it's worth considering whether
the computational cost is so bad compared with the extra lines of code
and opportunity for error.