Fix breakage of base construction
Commit 20aa225d545971d4da5912fbb07289b7d7aa8ad7 breaks the
base construction forms by use of incorrect form of None
check. "if result and result>=0 ..." is wrong because for
result==0 the expression "result" inside an if is treated as
False, leading to the entire offending condition to evaluate
to false. Fix this by converting the condition to the
correct "if result is not None ..."