PROTOTYPE: Add a base overview to the info screen
Give the player a quick insight into their bases and their status.  We
should how many bases can hold the singularity out of the total number
of bases.  In brackets, the player is informed of the bases under
construction, the bases complete but performing constructions as well
as the number of bases unable to maintain the singularity and also not
performing any item construction (usually you want them to build
CPUs).

The display is colored red on having only one (sustaining) base left,
yellow on non-sustaining bases that are not building anything and
white otherwise.

TODO:
 * Style the panel
 * De-hardcode the colors
 * Determine if the information are the correct ones

Signed-off-by: Niels Thykier <niels@thykier.net>