Refactor handling of region modifiers
Previously, we turned region modifiers into state data on the Location
objects.  As a consequence, savegames remembered the original region
modifier instead of getting the currently defined modifier.

This commit fixes this by serializing region information including
which modifier was given to a given location.  We store the modifier
"number" (list index) rather than the actual modifier.  Given this
number we can restore the modifier from the data file when loading a
savegame.

For older savegames, the modifier listed is reverse engineered from
the location's modifier state.

As a side-effect of this refactoring:

 * We also fix a bug under python3 where Australia would always be
   the location without a regional modifier.

 * Remove the data.reload_all_mutable() function.  Regions were the
   last case where we had mutable data outside the player object.
   (Though it had not been updated since mutable state for techs and
    events were moved to the player object)

Signed-off-by: Niels Thykier <niels@thykier.net>