Show correct total CPU per day when building CPUs for a base
This is first part of fixing #231 and ensures the player how much CPU
is gained from installing the CPU into a base with its *current*
modifiers.  If a network item has been constructed, it is also taken
into account (but it rarely matters as it is unlikely to be finished
when CPU is ordered).

Thanks: Wuzzy
Signed-off-by: Niels Thykier <niels@thykier.net>