Loadd dialog: Avoid crashing on attempting to "None" savegame
If you had 0 savegames (e.g. because you searched for a game that did
not exist) and activated "Load" (e.g. via Enter) the game would crash.
With this commit, we disable the "Load" button instead.

The "Delete" button is disabled for consistency (it did not trigger a
crash, but it seemed weird that only the "Load" button was grayed
out).

Signed-off-by: Niels Thykier <niels@thykier.net>