ResearchScreen: Special-case 0 available CPU
When people have 0 CPU (e.g. they put all their bases to sleep), then
both the slider and its "button" would take up the entire width of the
research screen.  However, if you have 1 CPU but 0 CPU for a given
danger level, affected research sliders do not.

With this commit, we special-case 0 CPU to take 10% of the width,
which matches how it looks with non-zero CPU but 0 CPU for the
relevant danger level.

Signed-off-by: Niels Thykier <niels@thykier.net>