Extract a StickyOnOffButton from ToggleButton
Extract a specialized "On/Off" button from ToggleButton.  The former
is basically a "group-less" version of the latter.  The specialization
enables us to support automatic translations of the button text.  This
in turn helps us simplify the Option code where we have a lot of
"On/Off" button where we need to translate their values.

Signed-off-by: Niels Thykier <niels@thykier.net>