Highlight locations with recent discoveries
After researching techs that give insight to detection and suspicion
values, highlight locations where the humans recently discovered a
base.

Partial detection insight shows last discovery and full detection
insight shows both locations (if they are the same, then a stronger
color is used)

The idea is based on an old play through guide, where the guide
suggested that you patched the game to highlight the locations.  It
serves as a good way of hinting to the player that there is a game
mechanic involving locations and base discovery.

Furthermore, it is an incentive to more experienced players on higher
difficulities to research those techs earlier as it makes the game
easier (by removing some micro management).

Signed-off-by: Niels Thykier <niels@thykier.net>