LogScreen: Rewrite how we replay messages
The replayed messages can easily become larger than the LogScreen,
which triggers warnings and make them display weirdly.  We "fix" this
by hiding the LogScreen and "rebase" the Message dialog on top of the
MapScreen.  When the dialog is closed, we unhide the LogScreen again.

This requires a bit of cumbersome code to do part of call_dialog's
work but make the dialog look much better.

Signed-off-by: Niels Thykier <niels@thykier.net>